backend/requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
supabase
python-dotenv
sqlalchemy
psycopg2-binary